- renamed the project, the solution and the folder - changed namespace - changed identifier and names in the info.plist
