+++ LocalNotifications/LocalNotifications/Assets.xcassets/AppIcons.appiconset/Icon-60@3x.png
