+++ LocalNotifications/LocalNotifications/Info.plist
@@ -3,9 +3,11 @@
 <plist version="1.0">
 <dict>
 	<key>CFBundleName</key>
-	<string>Notifications</string>
+	<string>LocalNotifications</string>
+	<key>CFBundleDisplayName</key>
+	<string>LocalNotification</string>
 	<key>CFBundleIdentifier</key>
-	<string>com.xamarin.notifications</string>
+	<string>com.xamarin.localnotifications</string>
 	<key>CFBundleShortVersionString</key>
 	<string>1.0</string>
 	<key>CFBundleVersion</key>
