+++ LocalNotifications/LocalNotifications/ViewController.cs
@@ -1,12 +1,11 @@
-﻿using System;
-using Foundation;
+﻿using Foundation;
+using System;
 using UIKit;
 
-namespace Notifications
+namespace LocalNotifications
 {
     public partial class ViewController : UIViewController
     {
-
         protected ViewController(IntPtr handle) : base(handle)
         {
             // Note: this .ctor should not contain any initialization logic.
@@ -47,4 +46,3 @@ namespace Notifications
         }
     }
 }
-
